Move traits::query::outlives_bounds::explicit_outlives_bounds to infer::outlives.
